Move fixtures from conftest to dedicated file
We want to reuse these in a conftest file for functional tests
so this is the easiest way of doing it.
